Move Admin API to api/admin.
Following on from how the primary API was lifted out of commands/serve/
into api/primary, we're doing the same thing with admin.
