+++ src/api/admin/events.rs
