+++ src/api/mod.rs
@@ -1,4 +1,5 @@
 // Copyright (c) 2025 Chris Corbyn <chris@zizq.io>
 // Licensed under the Business Source License 1.1. See LICENSE file for details.
 
+pub mod admin;
 pub mod primary;
